+++ CHANGELOG.md
@@ -4,10 +4,9 @@
 
 ### Features
 
-- Add `enableMemoryIntrospection` option to control memory introspection in crash reports (#7501)
-
 ### Fixes
 
+- Add `enableMemoryIntrospection` option to control memory introspection in crash reports, defaults to false (#7501)
 - Fix trace propagation for manually instrumented transactions when automatic performance tracing is disabled (#8522)
 - Remove x/y coordinates from UI breadcrumbs, as they are a potential security risk, for example leaking input on custom PIN code views (#8534)